integración de storage con crud/ integración de bootstrap y auth

+++ sistema/app/Http/Controllers/EmpleadoController.php
@@ -47,7 +47,8 @@ class EmpleadoController extends Controller
 
         Empleado::insert($datosEmpleado);
         
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+        return redirect('empleado')->with('mensaje','Empleado agregado con éxito');
     }
 
     /**
@@ -105,7 +106,12 @@ class EmpleadoController extends Controller
      */
     public function destroy($id)
     {
-        Empleado::destroy($id);
-        return redirect('empleado');
+        $empleado=Empleado::findOrFail($id);
+        if(Storage::delete('public/'.$empleado->Foto)){
+            Empleado::destroy($id);
+        }
+
+        
+        return redirect('empleado') ->with('mensaje','Empleado Borrado');
     }
 }
